+++ src/test/app/TestHostFunctions.h
@@ -22,7 +22,7 @@
 
 #include <xrpld/app/misc/AmendmentTable.h>
 #include <xrpld/app/tx/detail/NFTokenUtils.h>
-#include <xrpld/app/wasm/WasmHostFunc.h>
+#include <xrpld/app/wasm/HostFunc.h>
 #include <xrpld/app/wasm/WasmVM.h>
 #include <xrpld/ledger/detail/ApplyViewBase.h>
