+++ src/xrpld/app/tx/detail/Escrow.cpp
@@ -21,7 +21,7 @@
 #include <xrpld/app/misc/HashRouter.h>
 #include <xrpld/app/tx/detail/Escrow.h>
 #include <xrpld/app/tx/detail/MPTokenAuthorize.h>
-#include <xrpld/app/wasm/WasmHostFuncImpl.h>
+#include <xrpld/app/wasm/HostFuncImpl.h>
 #include <xrpld/app/wasm/WasmVM.h>
 #include <xrpld/conditions/Condition.h>
 #include <xrpld/conditions/Fulfillment.h>
