+++ src/xrpld/app/wasm/HostFuncImpl.h
@@ -20,7 +20,7 @@
 #pragma once
 
 #include <xrpld/app/tx/detail/ApplyContext.h>
-#include <xrpld/app/wasm/WasmHostFunc.h>
+#include <xrpld/app/wasm/HostFunc.h>
 
 namespace ripple {
 class WasmHostFunctionsImpl : public HostFunctions
