+++ src/xrpld/app/wasm/detail/WasmHostFuncImpl.cpp
@@ -19,7 +19,7 @@
 
 #include <xrpld/app/misc/AmendmentTable.h>
 #include <xrpld/app/tx/detail/NFTokenUtils.h>
-#include <xrpld/app/wasm/WasmHostFuncImpl.h>
+#include <xrpld/app/wasm/HostFuncImpl.h>
 
 #include <xrpl/protocol/STBitString.h>
 #include <xrpl/protocol/digest.h>
